requirements.txt
<<<<<<< Updated upstream
django 5.1.6
=======
django 
requests
python-dotenv
openai
requests
>>>>>>> Stashed changes
